+++ app/models/concerns/agilibox/pluck_distinct.rb
@@ -1,0 +1,13 @@
+module Agilibox::PluckDistinct
+  extend ActiveSupport::Concern
+
+  class_methods do
+    def pluck_distinct(column)
+      all
+        .reorder(column)
+        .select(column)
+        .distinct
+        .pluck(column)
+    end
+  end # class_methods
+end